feat: move notification helper methods to notifications_helper.rb
