+++ app/helpers/application_helper.rb
@@ -31,22 +31,6 @@ module ApplicationHelper
     !current_page?(notifications_path) && (authenticated_user_has_notifications? || unauthenticated_user_has_notifications?)
   end
 
-  def notification_count
-    if current_user.present?
-      current_user.active_unread_notifications.count
-    else
-      Notification.active_unauthenticated_notifications.count
-    end
-  end
-
-  def notification
-    if current_user.present?
-      current_user.newest_active_unread_notification
-    else
-      Notification.newest_active_unauthenticated_notification
-    end
-  end
-
 private
 
   def paginated_title(title, pagy)
